-Removed old uml diagram -fixed parent hierarchy for components and gameobjects -move circle on screen using arrow keys -lot changes in code: remove no longer used code now and later, added namespaces G1 for classes

+++ G1/CircleRenderer.h
@@ -11,7 +11,7 @@ namespace G1 {
 		void render(sf::RenderWindow*);
 	public:
 		CircleRenderer();
-		CircleRenderer(double radiusTemp, sf::Color);
+		CircleRenderer(float radiusTemp, sf::Color);
 		~CircleRenderer();
 	};
